Added the implementation of an engine with ARC cache
Please note that IANAL so I have not verified if this breaks any patents,
or checked the license of it.. I just added it to my repo to test it
out for fun (to see that the engine API works for other usecases as well).

This was posted to the mailing list by Limor Gavish lgavish@gmail.com
with the following comment:

I am working with Yuval and Nezer on the Memcached project

We have finished implementing ARC on the 1.6 branch (attached our files and the change in the makefile)
We have run some benchmarks and it seems to work quite similar to the original implementation (in the master branch).
We are running more tests and will upload when we have more results.