MB-18551: Synchronize delete buckets during shutdown
Earlier we just started the delete bucket threads and had
them run in parallel with the server shutdown. This could
lead to race condition for when resources would be released.

This patch runs through the list of buckets and then stops
each bucket in sequence and awaits for the delete to stop.

Change-Id: I5b4c7b64528f3cf19451f84956f7009439ee29e2